[#15] Refactor :Printer abstraction refactor(sponsor)
