+++ src/main/java/fr/socrates/domain/candidate/CandidateService.java
@@ -5,7 +5,5 @@ import java.util.List;
 public interface CandidateService {
     void add(Candidate candidate);
 
-    void print();
-
     List<Candidate> getRegisteredCandidates();
 }
